57 changes

+++ src/main/java/tech/atani/client/feature/guis/screens/mainmenu/atani/guis/AtaniServerListEntryNormal.java
@@ -147,7 +147,7 @@ public class AtaniServerListEntryNormal implements AtaniGuiListExtended.IGuiList
         else
         {
             k = 1;
-            l = (int)(Minecraft.getSystemTime() / 100L + (long)(slotIndex * 2) & 7L);
+            l = (int)(Minecraft.getSystemTime() / 100L + (slotIndex * 2L) & 7L);
 
             if (l > 4)
             {
